--- streamx-flink/streamx-flink-test/assembly/conf/hello.sql
@@ -1,2 +1,0 @@
-eNqtkDELweNqtkDELwjAQhff8ittOIQUrOEmHKkUL0qEGHCW2FynUpFwijAQhff8ittOIQUrOEmHKkUL0qEGHCW2FynUpFwi
-/n3TgkKd3R7H+94HJwTs6yJXBah8dyrAuyc3dA361hMsBIBZQVkpOab0m9ZwVnVZHQQs4VKq49TExllLTXCMkAG2Oug7WRwBZPfyyUCceIqtFjPcYIS3Qvz4B+5s+Id+Gvo4OuuJA8SLmxk89ZGIEmlSGWcNu8fsBSP8BvQiUI0=
