+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/streamx/test/tablesql/HelloFlinkSQL.scala
@@ -1,3 +1,23 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ * <p>
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package com.streamxhub.streamx.test.tablesql
 
 import com.streamxhub.streamx.flink.core.scala.FlinkStreamTable
@@ -5,6 +25,9 @@ import com.streamxhub.streamx.flink.core.scala.FlinkStreamTable
 object HelloFlinkSQL extends FlinkStreamTable {
 
   override def handle(): Unit = {
+    /**
+     * 一行胜千言
+     */
     context.sql("sql")
   }
 }
