+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/streamx/test/tablesql/KafkaTableApp.scala
@@ -1,3 +1,23 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ * <p>
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package com.streamxhub.streamx.test.tablesql
 
 import com.streamxhub.streamx.flink.core.scala.FlinkStreamTable
@@ -19,25 +39,6 @@ object KafkaTableApp extends FlinkStreamTable {
         "name" -> DataTypes.STRING()
       )
       .createTemporaryTable("kafka2Table")
-
-    val ds = context.$fromCollection(
-      List(
-        "flink,apapche flink",
-        "kafka,apapche kafka",
-        "spark,spark",
-        "zookeeper,apapche zookeeper",
-        "hadoop,apapche hadoop"
-      )
-    ).map(x => {
-      val array = x.split(",")
-      Entity(array.head, array.last)
-    })
-
-    context.createTemporaryView("kafkaSource", ds)
-
-    //kafka to table
-    val table1: Table = context.from("kafkaSource")
-    table1.>>[Entity].print("stream print==>")
 
     val table: Table = context.from("kafka2Table")
