--- streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/streamx/test/tablesql/MixinTableApp.scala
@@ -1,31 +1,0 @@
-package com.streamxhub.streamx.test.tablesql
-
-import com.streamxhub.streamx.flink.core.scala.FlinkStreamTable
-import org.apache.flink.api.scala._
-import org.apache.flink.table.api.Table
-
-object MixinTableApp extends FlinkStreamTable {
-
-  override def handle(): Unit = {
-
-    val mySource = context.$fromCollection(
-      List(
-        "kafka,apapche kafka",
-        "spark,spark",
-        "flink,apapche flink",
-        "zookeeper,apapche zookeeper",
-        "hadoop,apapche hadoop"
-      )
-    ).map(x => {
-      val array = x.split(",")
-      Entity(array.head, array.last)
-    })
-
-    context.createTemporaryView("mySource", mySource)
-
-    //kafka to table
-    val table1: Table = context.from("mySource")
-    table1.>>[Entity].print("stream print==>")
-
-  }
-}
