+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/streamx/test/tablesql/MixinTableDataStreamApp.scala
@@ -18,44 +18,34 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.streamxhub.streamx.test.stream
+package com.streamxhub.streamx.test.tablesql
 
-import com.streamxhub.streamx.flink.core.scala.sink.ClickHouseSink
-import com.streamxhub.streamx.flink.core.scala.{FlinkStreaming, StreamingContext}
+import com.streamxhub.streamx.flink.core.scala.FlinkStreamTable
 import org.apache.flink.api.scala._
+import org.apache.flink.table.api.Table
 
-object ClickHouseSinkApp extends FlinkStreaming {
+object MixinTableDataStreamApp extends FlinkStreamTable {
 
   override def handle(): Unit = {
-    val createTable =
-      """
-        |create TABLE test.orders(
-        |userId UInt16,
-        |orderId UInt16,
-        |siteId UInt8,
-        |cityId UInt8,
-        |orderStatus UInt8,
-        |price Float64,
-        |quantity UInt8,
-        |timestamp UInt16
-        |)ENGINE = TinyLog;
-        |""".stripMargin
 
-    println(createTable)
-
-    val source = context.addSource(new TestSource)
-
-    var index = 0
-    val httpDs = source.map(x => {
-      index += 1
-      s"""http://www.qq.com?id=$index"""
+    val mySource = context.$fromCollection(
+      List(
+        "kafka,apapche kafka",
+        "spark,spark",
+        "flink,apapche flink",
+        "zookeeper,apapche zookeeper",
+        "hadoop,apapche hadoop"
+      )
+    ).map(x => {
+      val array = x.split(",")
+      Entity(array.head, array.last)
     })
 
-    //HttpSink(context).getSink(httpDs).setParallelism(1)
+    context.createTemporaryView("mySource", mySource)
 
-    ClickHouseSink().sink[TestEntity](source, "test.orders")(x => {
-      s"${x.userId},${x.siteId}"
-    }).setParallelism(1)
+    //kafka to table
+    val table1: Table = context.from("mySource")
+    table1.>>[Entity].print("stream print==>")
+
   }
-
 }
